requirements.txt
prefect==2.10.5
dbt-core==1.4.6
dbt-bigquery==1.4.3
pyarrow==11.0.0
fastparquet==2023.2.0
python_dotenv==1.0.0
sqlalchemy==1.4.47
prefect-gcp==0.4.1
pandas-gbq==0.19.1